+++ gslab_scons/tests/_side_effects.py
@@ -229,153 +229,3 @@ def dot_git_open_side_effect(repo   = 'repo',
         return file_object
 
     return open_side_effect
-
-
-#== Side effects for testing create_size_dictionary() ===
-# These functions mock a directory containing files of various sizes
-# and a system that recognises no directory other than that one.
-def isdir_side_effect(*args, **kwargs):
-    '''
-    Mock os.path.isdir() so that it only recognises ./test_files/
-    as an existing directory.
-    '''
-    path = args[0]
-    if not isinstance(path, str):
-        raise TypeError('coercing to Unicode: need string or buffer, '
-                        '%s found' % type(path))
-    return path == 'test_files'
-
-
-def walk_side_effect(*args, **kwargs):
-    '''Mock os.walk() for a mock directory called ./test_files/.'''
-    path = args[0]
-
-    if path != 'test_files':
-        raise StopIteration
-
-    roots       = ['test_files',      'test_files/size_test']
-    directories = [['size_test'],     []]
-    files       = [['root_file.txt'], ['test.txt', 'test.pdf']]
-
-    for i in range(len(roots)):
-        yield (roots[i], directories[i], files[i])
-
-
-def getsize_side_effect(*args, **kwargs):
-    '''
-    Mock os.path.getsize() to return mock sizes of files in our
-    mock ./test_files/ directory.
-    '''
-    path = args[0]
-    if path == 'test_files/root_file.txt':
-        size = 100
-    elif path == 'test_files/size_test/test.txt':
-        size = 200
-    elif path == 'test_files/size_test/test.pdf':
-        size = 1000
-    else:
-        raise OSError("[Errno 2] No such file or directory: '%s'" % path)
-
-    return size
-
-
-# Define the mock file structure for testing list_ignored_files()
-struct = {'.': ['untracked.txt', 'make.log', 'make.py'],
-         'raw': ['large_file.txt', 'small_file.txt'], 
-         'release': ['output.txt', '.DS_Store'],
-         'release/subdir': ['ignored.txt']}
-
-
-def check_ignored_side_effect(status):
-
-    def effect(*args, **kwargs):
-        '''Mock subprcess.check_output() for testing list_ignored_files()'''
-    
-        # Define mock messages from "git status --ignored"
-        # i) Some files are ignored
-        standard = \
-            ['On branch testing\n',
-             'Your branch is up-to-date with \'origin/testing\'.\n',
-             'Changes not staged for commit:\n',
-             '  (use "git add/rm <file>..." to update what will be committed)\n',
-             '  (use "git checkout -- <file>..." to discard changes in working directory)\n',
-             '\n',
-             '\tmodified:   make.log\n',
-             '\n',
-             'Untracked files:\n',
-             '  (use "git add <file>..." to include in what will be committed)\n',
-             '\n',
-             '\tuntracked.txt',
-             '\n',
-             'Ignored files:\n',
-             '  (use "git add -f <file>..." to include in what will be committed)\n',
-             '\n',
-             '\traw/large_file.txt\n',
-             '\trelease/.DS_Store\n',
-             '\trelease/subdir/'
-             '\n',
-             '\n',
-             'It took 2.44 seconds to enumerate untracked files. \'status -uno\'\n',
-             'may speed it up, but you have to be careful not to forget to add\n',
-             'new files yourself (see \'git help status\').\n',
-             'no changes added to commit (use "git add" and/or "git commit -a")\n']
-        
-        # ii) No files are ignored
-        none_ignored = \
-            ['On branch issue59-size_warnings\n',
-             'Your branch is up-to-date with \'origin/issue59-size_warnings\'.\n',
-             'nothing to commit, working tree clean\n']
-
-        if status == 'none_ignored':
-            message = ''.join(none_ignored)
-        else:
-            message = ''.join(standard)
-
-        command = args[0]
-        if 'shell' in kwargs.keys():
-            if kwargs['shell'] and (command == 'git status --ignored'):
-                return message
-    
-        return None
-
-    return effect
-
-
-def walk_ignored_side_effect(*args, **kwargs):
-    '''Mock os.walk() for testing list_ignored_files()'''
-    path = args[0]
-    path = os.path.relpath(path)
-
-    if path not in struct.keys():
-        raise StopIteration
-
-    if path == 'release':
-        subdirs = ['subdir']
-    else:
-        subdirs = []
-
-    yield (path, subdirs, struct[path])
- 
-
-def isdir_ignored_side_effect(*args, **kwargs):
-    path = args[0]
-    if path == '':
-        return False
-        
-    isdir = (os.path.relpath(path) in struct.keys())
-    return isdir
-
-
-def isfile_ignore_side_effect(*args, **kwargs):
-    path = args[0]
-    if path == '':
-        return False
-
-    file_list = []
-
-    for k in struct.keys():
-        file_list += [os.path.join(k, f) for f in struct[k]]
-
-    isfile = (os.path.relpath(path) in map(os.path.relpath, file_list))
-    return isfile
-
